Allow passing fewer than 35 or 25 enrollment test answers (#90)
In #88 (that closes #89) we determine a level based on number of the
correct answers. This is not optimal as the user should be allowed to stop the assessment prematurely (which would be an equivalent of replying "I don't know" to all the remaining questions)
and still get the level calculated correctly, since the level depends
not only on the number of correct answers, but also on total number of
questions.

Hence, the client can pass to the backend another item in `data`: number
of questions. This should be very easy because the client just has to
count the questions as it receives them from the backend, and store this
number. Then the client passes this number back to the backend along
with the answers, thus allowing the backend to calculate the level
correctly without additional DB hits required to get to the
`EnrollmentTest` and count its questions.

Optionally, if client doesn't allow the user to interrupt the test, it
can just send the number of answers as number of questions.